Refactored the parsing of data into its own js file.
